requirements.txt
# ── 딥러닝 프레임워크 ───────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── Segmentation 모델 ───────────────────────────────────────
segmentation-models-pytorch>=0.3.3

# ── 데이터 증강 ─────────────────────────────────────────────
albumentations>=1.3.0

# ── 이미지 처리 ─────────────────────────────────────────────
opencv-python>=4.7.0
Pillow>=9.0.0

# ── 수치 연산 ───────────────────────────────────────────────
numpy>=1.23.0

# ── 시각화 ──────────────────────────────────────────────────
matplotlib>=3.6.0

# ── 진행률 표시 ─────────────────────────────────────────────
tqdm>=4.64.0
